fixed-lint-issue

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,7 +1,7 @@
 #include "function_pointers.h"
 
 /**
-  * array_iterator - executes a function given as a parameter on each element of an array.
+  * array_iterator - executes function given as a parameter on each element of an array.
   * @array: array passed as param
   * @size: size
   * @action: action (function) to execute
@@ -10,7 +10,7 @@
   */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i = 0;
+	unsigned int i = 0;
 
 	if (array != NULL && action != NULL && size > 0)
 	{
